requirements.txt
selenium==4.16.0
html2image==2.0.4.3
python-telegram-bot==20.7